# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refyne API",
    "version": "1.0.0",
    "description": "LLM-powered web extraction API that transforms unstructured websites into clean, typed JSON."
  },
  "servers": [
    {
      "url": "https://api.refyne.uk",
      "description": "Production API"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
    "/api/v1/extract": {
      "post": {
        "operationId": "extract",
        "tags": ["Extraction"],
        "summary": "Extract data from URL",
        "description": "Extract structured data from a single webpage using an LLM. Define a schema specifying the data structure you want, and the API returns JSON matching that schema.",
        "security": [{"Authorization": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic extraction",
                  "value": {
                    "url": "https://example.com/product/123",
                    "schema": {
                      "name": "string",
                      "price": "number",
                      "description": "string"
                    }
                  }
                },
                "with_llm_config": {
                  "summary": "With custom LLM provider",
                  "value": {
                    "url": "https://example.com/product/123",
                    "schema": {
                      "name": "string",
                      "price": "number"
                    },
                    "llm_config": {
                      "provider": "anthropic",
                      "model": "claude-3-5-sonnet-20241022"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful extraction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "description": "Include your token in the Authorization header."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "required": ["url", "schema"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "URL to extract data from"
          },
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "Schema defining the data structure to extract (JSON or YAML)"
          },
          "fetch_mode": {
            "type": "string",
            "enum": ["auto", "static", "dynamic"],
            "default": "auto",
            "description": "Fetch mode: auto (detect), static (fast), or dynamic (JavaScript rendering)"
          },
          "llm_config": {
            "$ref": "#/components/schemas/LlmConfig"
          }
        }
      },
      "ExtractResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Extracted data matching the schema"
          },
          "url": {
            "type": "string",
            "description": "URL that was extracted"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the page was fetched"
          },
          "usage": {
            "$ref": "#/components/schemas/TokenUsage"
          },
          "metadata": {
            "$ref": "#/components/schemas/ExtractionMetadata"
          }
        }
      },
      "TokenUsage": {
        "type": "object",
        "properties": {
          "input_tokens": {
            "type": "integer",
            "description": "Number of input tokens used"
          },
          "output_tokens": {
            "type": "integer",
            "description": "Number of output tokens used"
          },
          "cost_usd": {
            "type": "number",
            "description": "Total USD cost charged"
          },
          "llm_cost_usd": {
            "type": "number",
            "description": "Actual LLM cost from provider"
          },
          "is_byok": {
            "type": "boolean",
            "description": "True if user's own API key was used"
          }
        }
      },
      "ExtractionMetadata": {
        "type": "object",
        "properties": {
          "fetch_duration_ms": {
            "type": "integer",
            "description": "Time to fetch the page in milliseconds"
          },
          "extract_duration_ms": {
            "type": "integer",
            "description": "Time to extract data in milliseconds"
          },
          "model": {
            "type": "string",
            "description": "Model used for extraction"
          },
          "provider": {
            "type": "string",
            "description": "LLM provider used"
          }
        }
      },
      "LlmConfig": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["anthropic", "openai", "openrouter", "ollama", "credits"],
            "description": "LLM provider"
          },
          "api_key": {
            "type": "string",
            "description": "API key for the provider (BYOK)"
          },
          "base_url": {
            "type": "string",
            "description": "Custom base URL (for Ollama or self-hosted)"
          },
          "model": {
            "type": "string",
            "description": "Model to use"
          }
        }
      },
...
    }
  }
}